pyproject.toml
[project]
name = "compulsor"
version = "1.0.0"
description = "Post Community Team Pulse Reports"
authors = [
  { name = "Philipp Kewisch", email = "code@kewis.ch" }
]
requires-python = ">=3.6"
dependencies = [
  "jira",
  "jira2markdown",
  "pyyaml",
  "click",
  "pydiscourse",
  "pyspellchecker"
]
[project.urls]
homepage = "https://github.com/kewisch/compulsor"

[project.optional-dependencies]
dev = [
  "black",
  "isort",
  "flake8",
  "Flake8-pyproject",
  "tox"
]

[project.scripts]
compulsor = "compulsor.compulsor:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.flake8]
ignore = ["E203", "E266", "E501", "W503"]
max-line-length = 100
max-complexity = 18
select = ["B","C","E","F","W","T4"]


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py3

[testenv]
deps =
  black
  isort
  flake8
  Flake8-pyproject
commands =
 black compulsor
 isort compulsor
 flake8 compulsor
"""
